modification login

+++ base/manadiobase.sql
@@ -9,7 +9,7 @@ BEGIN
     
     -- Supprimer toutes les données de toutes les tables, sauf celle de l'utilisateur
     FOR r IN (SELECT tablename FROM pg_tables WHERE schemaname = current_schema()) LOOP
-        IF r.tablename <> 'utilisateur' THEN
+        IF r.tablename <> 'utilisateur' and r.tablename <> 'sexe' THEN
             EXECUTE 'DELETE FROM ' || quote_ident(r.tablename);
         END IF;
     END LOOP;
